% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Com2seq_fun.R
\name{Com2seq}
\alias{Com2seq}
\title{Com2seq: combining two sequencing datasets (e.g., from 16S marker-gene and shotgun metagenomic sequencing) for integrative analysis of differential abundance at both the genus and community levels}
\usage{
Com2seq(
  table1,
  table2,
  Y1,
  Y2,
  C1 = NULL,
  C2 = NULL,
  fdr.nominal = 0.2,
  seed = NULL,
  n.perm.max = NULL,
  n.rej.stop = 100,
  n.cores = 4,
  filter.thresh = 0.2,
  verbose = TRUE
)
}
\arguments{
\item{table1}{taxa count (relative abundance) table agglomerated at the genus level. The rows correspond to samples and the columns correspond to genera.}

\item{table2}{taxa count (relative abundance) table agglomerated at the genus level. The rows correspond to samples and the columns correspond to genera.}

\item{Y1}{trait of interest, with samples ordered in the same way as in \code{table1}. It can be an array, matrix, or data frame; a factor should be represented by the corresponding design matrix. All components in a matrix or data frame will be tested for microbial association jointly.}

\item{Y2}{trait of interest, with samples ordered in the same way as in \code{table2}. See other requirements in \code{Y1}.}

\item{C1}{other covariates, with samples ordered in the same way as in \code{table1}. See other requirements in \code{Y1}.}

\item{C2}{other covariates, with samples ordered in the same way as in \code{table2}. See other requirements in \code{Y1}.}

\item{fdr.nominal}{the nominal FDR. The default is 0.2.}

\item{seed}{a user-supplied integer seed for the random number generator in the
permutation procedure. The default is NULL, which means that an integer seed will be
generated internally and randomly. In either case, the integer seed will be stored
in the output object in case the user wants to reproduce the permutation replicates.}

\item{n.perm.max}{the maximum number of permutations. The default is NULL, in which case \code{n.taxa} * \code{n.rej.stop} * (1/\code{fdr.nominal})
is used where \code{n.taxa} is the total number of taxa (that have non-zero counts in at least one sample).}

\item{n.rej.stop}{the minimum number of rejections (i.e., the permutation test
statistic exceeds the observed test statistic) to obtain before stopping the permutation procedure. The
default is 100.}

\item{n.cores}{the number of cores to be used for parallel computing. The default is 1.}

\item{filter.thresh}{a real value between 0 and 1; OTUs present in fewer than \code{filter.thresh} * 100\% samples are filtered out. The default is 0.2.}

\item{verbose}{a logical value indicating whether to generate verbose output during the permutation process. Default is TRUE.}
}
\value{
A list consisting of
\itemize{
  \item effect.size.wc - effect size at each taxon, i.e., beta_j,1 - median_j'=1,...J (beta_j',1), based on the count weight
  \item p.taxa.wc - p-values by the taxon-specific tests based on the count weight
  \item q.taxa.wc - q-values (adjusted p-values by the BH procedure) by the taxon-specific tests based on the count weight
  \item detected.taxa.wc - detected taxa (using the column names of the taxon table) at the nominal FDR by the taxon-specific tests based on the count weight
  \item p.global.wc - p-value by the global test based on the count weight
  \item effect.size.we - effect size at each taxon, i.e., beta_j,1 - median_j'=1,...J(beta_j',1), based on the equal weight
  \item p.taxa.we - p-values by the taxon-specific tests based on the equal weight
  \item q.taxa.we - q-values (adjusted p-values by the BH procedure) by the taxon-specific tests based on the equal weight
  \item detected.taxa.we - detected taxa (using the column names of the taxon table) at the nominal FDR by the taxon-specific tests based on the equal weight
  \item p.global.we - p-value by the global test based on the equal weight
  \item p.taxa.omni - p-values by the taxon-specific omnibus tests 
  \item q.taxa.omni - q-values (adjusted p-values by the BH procedure) by the taxon-specific omnibus tests 
  \item detected.taxa.omni - detected taxa (using the column names of the taxon table) at the nominal FDR by the taxon-specific omnibus tests 
  \item p.global.omni - p-value by the global omnibus test 
  \item n.perm.completed - number of permutations completed
  \item ref.taxon - the selected reference taxon
  \item seed - the seed used to generate the permutation replicates
}
}
\description{
This function allows you to test
(1). whether any genus is associated with the trait of interest with FDR control, based on the log ratio of relative abundances between pairs of taxa, and
(2). whehter the whole community is associated with the trait (a global test).
The tests accommodate continuous, discrete (binary, categorical), and multivariate traits, and allows adjustment of confounders.
}
\details{
This function uses a sequential stopping criterion (similar to that of Sandve et al. 2011) for the permutation procedure,
which stops when all taxon-level tests have either reached the pre-specified
number (default 100) of rejections or yielded a q-value (by the Benjamini-Hochberg [BH] procedure) that is below the
nominal FDR level (default 0.2). The permutation procedure is always terminated if a pre-specified maximum number (see description of \code{n.perm.max}) of
permutations have been generated. The global test is based on all permutation replicates when the procedure stops/terminates.
}
\examples{
library(LOCOM)
data("otu.table.1")
data("otu.table.2")
data("meta.1")
data("meta.2")

# preparing Y1, Y2, C1, C2
Y1 <- cbind(1*(meta.1[,"Y"]==1), 1*(meta.1[,"Y"]==2)) # get design matrix for a three-level factor
Y2 <- cbind(1*(meta.2[,"Y"]==1), 1*(meta.2[,"Y"]==2))
C1 <- meta.1[,"C"] 
C2 <- meta.2[,"C"] 

# restricting to a subset of samples and OTUs for speed
sub.sam.1 <- c(31:60)
sub.sam.2 <- c(1:20, 51:60)
otu.table.1 <- otu.table.1[sub.sam.1, 1:70]
otu.table.2 <- otu.table.2[sub.sam.2, 1:70]
Y1 <- Y1[sub.sam.1,]
C1 <- C1[sub.sam.1]
Y2 <- Y2[sub.sam.2,]
C2 <- C2[sub.sam.2]

# running Com2seq
res.Com2seq <- Com2seq(table1 = otu.table.1, table2 = otu.table.2, 
                       Y1 = Y1, Y2 = Y2, C1 = C1, C2 = C2, 
                       seed = 123, n.cores = 1, n.perm.max = 1000)
res.Com2seq$p.global.omni 
res.Com2seq$detected.taxa.omni 
}
