% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.1.R
\docType{data}
\name{meta.1}
\alias{meta.1}
\title{Metadata for samples in otu.table.1}
\format{
A data frame with 80 observations on 2 variables.
}
\usage{
data("meta.1")
}
\description{
Metadata for samples in otu.table.1
}
\examples{
data(meta.1)
}
\keyword{datasets}
