% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.2.R
\docType{data}
\name{meta.2}
\alias{meta.2}
\title{Metadata for samples in otu.table.2}
\format{
A data frame with 60 observations on 2 variables.
}
\usage{
data("meta.2")
}
\description{
Metadata for samples in otu.table.2
}
\examples{
data(meta.2)
}
\keyword{datasets}
