% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu.table.1.R
\docType{data}
\name{otu.table.1}
\alias{otu.table.1}
\title{Simulated OTU count table}
\format{
A data frame with 80 observations on 677 variables.
}
\usage{
data("otu.table.1")
}
\description{
This table contains 80 subjects with 677 OTUs.
}
\examples{
data(otu.table.1)
}
\keyword{datasets}
