% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu.table.2.R
\docType{data}
\name{otu.table.2}
\alias{otu.table.2}
\title{Simulated OTU count table}
\format{
A data frame with 80 observations on 695 variables.
}
\usage{
data("otu.table.2")
}
\description{
This table contains 60 subjects with 695 OTUs.
}
\examples{
data(otu.table.2)
}
\keyword{datasets}
