% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throat.otu.taxonomy.R
\docType{data}
\name{throat.otu.taxonomy}
\alias{throat.otu.taxonomy}
\title{Taxonomy names for OTUs from 16S sequencing of the throat microbiome samples}
\format{
A vector with 5683 taxonomy names
}
\source{
Charlson ES, Chen J, Custers-Allen R, Bittinger K, Li H, et al. (2010) 
Disordered Microbial Communities in the Upper Respiratory Tract of Cigarette Smokers. PLoS ONE 5(12): e15216.
}
\usage{
data("throat.otu.taxonomy")
}
\description{
This file contains 5683 taxonomy names.
}
\examples{
data(throat.otu.taxonomy)
}
\references{
R package "GUniFrac"
}
\keyword{datasets}
